refactor PacketHeaderConstraintToFlowHelper to IpFieldExtractorContext
